Don't force the download of the latest version of an index if comments are not activated
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/apps/Thaw@14943 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/thaw/plugins/index/Index.java
@@ -954,7 +954,8 @@ public class Index extends Observable implements MutableTreeNode,
 						parser.loadXML(path);
 
 
-						if (!fetchingNegRev && mustFetchNegRev) {
+						if (!fetchingNegRev && mustFetchNegRev
+						    && getCommentPublicKey() != null) {
 							final java.io.File fl = new java.io.File(path);
 							fl.delete();
